pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-lecture-platform"
version = "0.1.0"
description = "웹 기반 슬라이드형 강의안 플랫폼"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Flask==2.3.3",
    "Flask-CORS==4.0.0",
    "Jinja2==3.1.2",
    "Werkzeug==2.3.7",
    "python-dotenv==1.0.0",
    "Pillow==10.0.1",
    "markdown==3.5.1",
    "PyYAML==6.0.1",
    "gunicorn==21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.991",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true 